Level-5: Handle Errors
- handle errors
- cleaned up code by creating new file for each class
