+++ src/main/java/duke/Task.java
@@ -1,0 +1,53 @@
+package duke;
+
+import java.util.ArrayList;
+
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public void markAsDone() {
+        this.isDone = true;
+    }
+
+    public void markAsUnDone() {
+        this.isDone = false;
+    }
+
+    public String toString() {
+        return  "[" + this.getStatusIcon() + "] " + this.description;
+    }
+
+    public static void displayList(ArrayList<Task> array) {
+        System.out.println("    ────────────── ⋆⋅☆⋅⋆ ───────────────");
+        for(int j=0; j<array.size(); j++) {
+            System.out.println("      "+(j+1) +"."+ array.get(j).toString());
+        }
+        System.out.println("    ────────────── ⋆⋅☆⋅⋆ ───────────────");
+    }
+
+    public static void unmarkTask(ArrayList<Task> array, String[] splitInput) {
+        array.get((Integer.parseInt(splitInput[1])-1)).markAsUnDone();
+        System.out.println("    ────────────── ⋆⋅☆⋅⋆ ───────────────");
+        System.out.println("      OK, I've marked this task as not done yet:");
+        System.out.println("      "+ array.get((Integer.parseInt(splitInput[1])-1)).toString());
+        System.out.println("    ────────────── ⋆⋅☆⋅⋆ ───────────────");
+    }
+
+    public static void markTask(ArrayList<Task> array, String[] splitInput) {
+        array.get((Integer.parseInt(splitInput[1])-1)).markAsDone();
+        System.out.println("    ────────────── ⋆⋅☆⋅⋆ ───────────────");
+        System.out.println("      Nice! I've marked this task as done:");
+        System.out.println("      "+ array.get((Integer.parseInt(splitInput[1])-1)).toString());
+        System.out.println("    ────────────── ⋆⋅☆⋅⋆ ───────────────");
+    }
+}
