+++ src/main/java/duke/Todo.java
@@ -1,0 +1,45 @@
+package duke;
+
+import util.*;
+import java.util.ArrayList;
+
+public class Todo extends Task {
+
+    protected String duration;
+    static String divider = "    ═══*.·:·.☽✧    ✦    ✧☾.·:·.*═══";
+
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    public static void createTodoTask(ArrayList<Task> array, String[] splitInput) {
+        if (splitInput.length == 1 || splitInput[1].equals("")){
+            try {
+                throw new DukeException("todo");
+            } catch (Exception e) {
+                System.out.println(divider);
+                System.out.println(e.toString());
+                System.out.println(divider);
+            }
+        } else {
+            for(int j=2; j< splitInput.length; j++){
+                splitInput[1] = splitInput[1] + " " + splitInput[j];
+            }
+               
+            String desc = splitInput[1];
+            Todo t = new Todo(desc);
+            array.add(t);
+            System.out.println(divider);
+            System.out.println("     Got it. I've added this task:");
+            System.out.println("     " + t.toString());
+            System.out.println("     Now you have " + array.size() + " tasks in the list.");
+            System.out.println(divider);
+        }
+        
+    }
+}
