+++ text-ui-test/EXPECTED.TXT
@@ -1,7 +1,14 @@
-Hello from
- ____        _        
-|  _ \ _   _| | _____ 
-| | | | | | | |/ / _ \
-| |_| | |_| |   <  __/
-|____/ \__,_|_|\_\___|
+  ───── ･ ｡ﾟ☆: *.☽ .* :☆ﾟ. ───── 
+ Hello! I'm Broccoli the dinosaur 
+           <|°▿▿▿▿°|/ 
+      What can I do for you? 
+   ──── ･ ｡ﾟ☆: *.☽ .* :☆ﾟ. ────
+    ═══*.·:·.☽✧    ✦    ✧☾.·:·.*═══
+     Got it. I've added this task:
+     [T][ ] borrow book
+     Now you have 1 tasks in the list.
+    ═══*.·:·.☽✧    ✦    ✧☾.·:·.*═══
+    ──────── ⋅ ∙ ∘ ☽ ༓ ☾ ∘ ⋅ ⋅ ─────────
+      Bye. Hope to see you again soon!
 
+   ‧˚₊꒷꒦︶︶︶︶︶꒷꒦︶︶︶︶︶꒦꒷‧₊˚⊹
