+++ text-ui-test/input.txt
@@ -1,0 +1,2 @@
+todo borrow book
+bye
